Worker-ize encryption and decryption
