+++ package.json
@@ -200,6 +200,7 @@
     "webpack-dev-server": "^3.1.4",
     "webpackbar": "^2.6.1",
     "workbox-webpack-plugin": "^3.3.0",
+    "worker-loader": "2.0.0",
     "yargs": "^3.27.0"
   },
   "keywords": [
