+++ webpack.config.js
@@ -19,7 +19,10 @@ const isProd = process.env.NODE_ENV === 'production'
 const output = {
   filename: 'js/[name].[hash:8].js',
   chunkFilename: 'js/[name].[hash:8].chunk.js',
-  path: path.resolve(__dirname, 'build')
+  path: path.resolve(__dirname, 'build'),
+  publicPath: '/',
+  // Fix workers & HMR https://github.com/webpack/webpack/issues/6642
+  globalObject: 'self'
 }
 
 /**
@@ -31,6 +34,7 @@ const output = {
 if (isProd) {
   output.path = path.resolve(__dirname, 'build/static')
   output.publicPath = '/static/'
+  output.globalObject = undefined
 }
 
 /**
@@ -123,6 +127,10 @@ module.exports = {
             options: { minimize: true }
           }
         ]
+      },
+      {
+        test: /\.worker\.js$/,
+        use: 'worker-loader'
       }
     ]
   },
